feat: Complete service layer and controller implementation

+++ src/main/java/com/example/booking/trainticket/dto/TicketDto.java
@@ -18,12 +18,12 @@ import lombok.ToString;
 public class TicketDto {
 	private String pnr;		// Unique PNR for each ticket
     private BookingStatus bookingStatus;	// "CONFIRMED" or "WAITLISTED"
-    private int travellerId;// Associated Traveller ID
+    private Long travellerId;// Associated Traveller ID
     private String travellerName;
     private LocalDate departureDate;// Travel Date
     private String source;
     private String destination;
-    private int trainId;	// Associated Train ID
-    private int seatNumber;	// Seat Number (-1 if waitlisted)
+    private Long trainId;	// Associated Train ID
+    private Long seatNumber;	// Seat Number (-1 if waitlisted)
     private double fare;	// Ticket Fare
 }
